pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nbblackbox"
version = "0.9"
dependencies = [
    "django~=5.0",
    "mozilla-django-oidc~=3.0",
    "gunicorn~=21.2",
    "psycopg2-binary~=2.9",
    "whitenoise~=6.8.2",
    "django-csp~=3.8",
    "django-maintenance-mode~=0.22.0"
]